Splitted quote into multiple files, added comments
