+++ modules/Misc/quote/subcommands/quoteByID.js
@@ -1,0 +1,5 @@
+const sendQuote = require('../sendQuote.js');
+
+module.exports = (msg, args) => {
+    return sendQuote(msg, args[0], msg.channel);
+}
